Refs #37159 - Refactor content specific discoveries
